requirements.txt
asgiref==3.3.4
Django==3.1.11
django-cors-headers==3.4.0
djangorestframework==3.11.2
gunicorn==20.1.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
